impl Crud operations with product image

+++ src/main/java/com/paradisemidigama/SurfShop/controller/ProductController.java
@@ -3,35 +3,88 @@ package com.paradisemidigama.SurfShop.controller;
 import com.paradisemidigama.SurfShop.model.Product;
 import com.paradisemidigama.SurfShop.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api")
+@CrossOrigin
 public class ProductController {
 
     @Autowired
     private ProductService prodService;
 
     @GetMapping("/products")
-    public List<Product> getAllProducts(){
-        return prodService.getAllProducts();
+    public ResponseEntity<List<Product>> getAllProducts(){
+        return new ResponseEntity<>(prodService.getAllProducts(),HttpStatus.OK);
+    }
+
+    @GetMapping("/product/{id}")
+    public ResponseEntity<Product> getProduct(@PathVariable int id){
+
+        Product product = prodService.getProductById(id);
+
+        if(product != null){
+            return new ResponseEntity<>(product,HttpStatus.OK);
+
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping("/product")
-    public Product addProduct(@RequestBody Product product) {
-        System.out.println("Received Product: " + product);
-        return prodService.addProduct(product);
+    public ResponseEntity<?> addProduct(@RequestPart Product product,
+                                        @RequestPart MultipartFile imageFile){
+        try{
+            Product product1 = prodService.addProduct(product,imageFile);
+            return new ResponseEntity<>(product1,HttpStatus.CREATED);
+        }catch(Exception e){
+            return new ResponseEntity<>(e.getMessage(),HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @GetMapping("/product/{productId}/image")
+    public ResponseEntity<byte[]> getImageByProductId(@PathVariable int productId){
+        Product product = prodService.getProductById(productId);
+        byte[] imageFile = product.getImageData();
+
+        return ResponseEntity.ok()
+                .contentType(MediaType.valueOf(product.getImageType()))
+                .body(imageFile);
     }
 
     @PutMapping("/product/{id}")
-    public Product updateProduct(@PathVariable int id,@RequestBody Product product){
-        return prodService.updateProduct(id,product);
+    public ResponseEntity<String> updateProduct(@PathVariable int id, @RequestPart Product product,
+                                                @RequestPart MultipartFile imageFile){
+        Product product1 = null;
+        try {
+            product1 = prodService.updateProduct(id,product,imageFile);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        if(product1 != null){
+            return new ResponseEntity<>("Updated", HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>("Failed to Update",HttpStatus.BAD_REQUEST);
+        }
     }
 
-    @DeleteMapping("product/{id}")
-    public void deleteProduct(@PathVariable int id){
-        prodService.deleteProduct(id);
+    @DeleteMapping("/product/{id}")
+    public ResponseEntity<String> deleteProduct(@PathVariable int id){
+        Product product = prodService.getProductById(id);
+        if(product != null){
+            prodService.deleteProduct(id);
+            return new ResponseEntity<>("Deleted", HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>("Product Not Found",HttpStatus.NOT_FOUND);
+        }
     }
+
 }
